requirements.txt
coreapi == 2.3.3
django == 1.11
djangorestframework == 3.7.3
django-rest-swagger == 2.1.2
email_reply_parser == 0.5.9
google-api-python-client == 1.6.4
google-cloud-storage == 1.6.0
google-cloud-pubsub == 0.29.2
gunicorn == 19.7.1
markdown == 2.6.10
mkdocs
pygments == 2.2.0
requests == 2.18.4
skpy==0.8.2
telethon==0.19.1.6
djangorestframework_simplejwt == 3.2
django-oauth-toolkit == 0.11.0
scipy == 1.0.0
django-crontab == 0.7.1
mysqlclient == 1.3.12
mysql-connector-python == 8.0.6
-e git://github.com/Michotastico/WebWhatsAPI.git@master#egg=webwhatsapi